+++ NEWS.md
@@ -2,6 +2,7 @@
 * New functions is.metric() and is.synchronous().
 * Update of function trajectoryVariability().
 * New function variationDecomposition() for synchronous trajectories.
+* New function interpolateTrajectories() for interpolation to given times.
 
 # ecotraj 1.0.0
 * Trajectory data structures have been introduced.